extract common func from GamePage
