+++ modules/common/demo/index.ts
@@ -1,3 +1,4 @@
 export * from "./inflateDemo";
 export * from "./playDemo";
 export * from "./makeDemo";
+export * from "./demo2ui";
