+++ modules/common/empty/emptyAnim.ts
@@ -1,0 +1,7 @@
+import { AlgolAnimCompiled } from "../../types";
+
+export const emptyAnim: AlgolAnimCompiled = {
+  enterFrom: {},
+  exitTo: {},
+  ghosts: [],
+};
